🚚 refactor(web): rename features/automations to features/tasks
The directory backs the /agents/{id}/tasks/* pages; "automations" was the last
remnant of the old IA. Pure rename — git mv the folder, repoint the seven
external importers, and drop the dead nav.automations i18n keys. Internal
imports are relative and untouched; the automations.* message-key namespace
stays (it names keys, not paths).

Refs: #509
