+++ web/src/features/tasks/DetailShell.tsx
