+++ web/src/routes/_app/agents.$agentId/tasks/goals.lazy.tsx
@@ -1,5 +1,5 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
-import { GoalsPage } from "@/features/automations/GoalsPage";
+import { GoalsPage } from "@/features/tasks/GoalsPage";
 
 export const Route = createLazyFileRoute("/_app/agents/$agentId/tasks/goals")({
   component: GoalsPage,
