+++ web/src/routes/_app/agents.$agentId/tasks/index.lazy.tsx
@@ -1,5 +1,5 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
-import { OverviewPage } from "@/features/automations/OverviewPage";
+import { OverviewPage } from "@/features/tasks/OverviewPage";
 
 export const Route = createLazyFileRoute("/_app/agents/$agentId/tasks/")({
   component: OverviewPage,
